requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
streamlit==1.28.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dateutil==2.8.2
regex==2023.10.3 